feat: changed test name
